+++ include/bitreverse.hpp
@@ -4,6 +4,6 @@
 #include <typedefs.hpp>
 
 auto bit_reverse_copy(vcpx& v) -> void;
-auto next_power_of_2(long unsigned int n) -> long unsigned int;
+auto next_power_of_2(size_t n) -> size_t;
 
 #endif
